[add] CompanyInfoとDetailのバグ修正 (#37)
