+++ backend/domain/area.go
@@ -12,8 +12,8 @@ type Area struct {
 	City          string `json:"city" gorm:"not null"`
 	AddressNumber int    `json:"address_number"`
 	BuildingName  string `json:"building_name"`
-	//has one
-	Detail Detail `json:"detail"`
+	// //has one
+	// Detail Detail `json:"detail"`
 	//time stamp
 	UpdatedAt time.Time `json:"updated_at"`
 	CreatedAt time.Time `json:"created_at"`
